# Edit sunpy.io._cdf.read_cdf to avoid Pandas PerformanceWarning: DataFrame is highly fragmented (fix #7246)