--- a/changelog/7247.bugfix.rst
+++ b/changelog/7247.bugfix.rst
@@ -1,0 +1,1 @@
+Improved code when loading CDF files to improve performance and avoid raising of pandas performance warnings.