## PR Description

As described in #7246, when reading cdf files with a lot of variables, for each variable `PerformanceWarning: DataFrame is highly fragmented.` is raised by Pandas, as each variable is added to a DataFrame using `df[var_key] = data`.

This PR changes the function `sunpy.io._cdf.read_cdf` so that it gathers the data for all variables in a dictionary throughout the for-loop (over each variable), and in the end concat it to the desired DataFrame, which according to some comments online is memory wise more efficient.

This gets rid of the PerformanceWarnings, and it seems to run a little bit faster on my machine, for the above example using `%timeit -n1 -r3` it's `7.69 s ± 10.8 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)` vs. `6.79 s ± 67.5 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)`. 

I don't have any official statement that this fix is really better or the way to go, but as it removes annoying warnings and seems to be not slower, maybe it's worthwhile to implement it.

Fixes #7246